backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["dhythm <yuta.okada.20@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.3"
djangorestframework = "^3.14.0"
pyjwt = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
